コードの整理と型キャスト追加、エフェクト制御の変更
Particle.cpp: `scaleMultiplier`の定数定義をコメントアウトし、スケール調整を無効化。
Particle.h: `SetParticleShape`関数で形状設定を無効化。
Line.cpp: `DrawInstanced`と`vertexBufferView_.SizeInBytes`に型キャストを追加。
LineManager.cpp: `_DEBUG`条件削除、未使用変数を保持。
imgui_impl_dx12.cpp: `hr`を未使用のまま記述。
main.cpp: `WinMain`関数で未使用引数を保持。
DebugScene.cpp: `objHammer_`と`objMicrowave_`の描画をコメントアウト。
DebugScene.h: エフェクト制御変数を変更し、ダイナミックヒットとアーク溶接エフェクトを有効化。

+++ engine/3d/Line.cpp
@@ -101,7 +101,7 @@ void Line::Draw() {
 	commandList->IASetVertexBuffers(0, 1, &vertexBufferView_);
 	//========================================
 	// 描画
-	commandList->DrawInstanced(vertices_.size(), 1, 0, 0);
+	commandList->DrawInstanced(static_cast<UINT>( vertices_.size() ), 1, 0, 0);
 	// NOTE:描画した後はラインをクリアするのを忘れるな
 }
 
@@ -151,7 +151,7 @@ void Line::CreateVertexBuffer() {
 	// バーテックスバッファビューの設定
     vertexBufferView_.BufferLocation = vertexBuffer_->GetGPUVirtualAddress();
 	// バイトサイズ
-    vertexBufferView_.SizeInBytes = bufferSize;
+	vertexBufferView_.SizeInBytes = static_cast<UINT>( bufferSize );
 	// ストライド
     vertexBufferView_.StrideInBytes = sizeof(LineVertex);
 }
